expectations: fix default Result value.
Assume that all entries in our expectation files are EXEC_FAILED
by default. That has been the default assumption in practice and
tests that are expected to succeed are already explicitly tagged
as SUCCESS.

Note that the "isFailure" boolean (based on the "failure:" element)
is largely orthogonal and can safely be removed: It's only used by
4 legacy rules. Moreover, it's main usage seems to be from a
private method with no caller.

bug: 22062343
Change-Id: I50afff7624f0ad4c1d29ead17dc1eb5312bda413