Fix some of the more obvious usages of private APIs in CTS tests
Replaces some usages of com.google.android.collect.Sets, some unused
import of framework-private classes.

Uses a hack to generate a MenuBuilder in MenuInflaterTest since the
entire class is framework-private and it's a huge pain to build an
equivalent MockMenu.

Uses Relative.LayoutParams.resolveLayoutDirection() to emulate the
behavior of getRules(int) which is framework-private.

Bug: 13281485
Change-Id: Id8b94d73b06a9a8e14a8aac56fc06de06612e7c1