Test that SysV IPC has been removed from the kernel
Returns true if SysV IPC has been removed.

System V IPCs are not compliant with Android's application lifecycle
because allocated resources are not freed by the low memory killer.
This lead to global kernel resource leakage.

For example, there is no way to automatically release a SysV
semaphore allocated in the kernel when:
- a buggy or malicious process exits
- a non-buggy and non-malicious process crashes or is explicitly killed.

Killing processes automatically to make room for new ones is an
important part of Android's application lifecycle implementation.
This means that, even assuming only non-buggy and non-malicious code,
it is very likely that over time, the kernel global tables used to
implement SysV IPCs will fill up.

Bug: 19340456
Bug: 22300191
Change-Id: I348dfdc58750b6318bc964a212f5f8b0015b72dd