Update settings provider - insert in secure settings test
The settings provider fails fast if the input for the insert method
is not valid. The test is trying to insert a null key which is not
a valid key. Note that in this case no mutation is performed. To really
tests the provider protection we are inserting a valid key, value pair.

bug:21619456

Change-Id: Iadd9208deda1fa44eb2fc8a76347a19b417e0010