Fix cts test in DumpsysHostTest#testBatteryStatsOutput
The output in BatteryStats is comma separated. Some of the values
can be the empty string. When testing for the number of entries,
We call String.split(str) and count the length of the resulting array.
Unfortunately, when an empty string is the last value in the comma
separated list, it is dropped.

Using String.split(str, -1) retains the trailing empty string.

Bug:22807424
Change-Id: I666f27614122bc962351e4fca681d838fd656df8