Add the ICU tests to CTS.
The external ICU library has it's own testing framework which
is not JUnit. This change introduces a new test runner for
running ICU tests and reporting on the results. This includes
resolving the correct package names to allow tests to be
specified using the fully qualified ICU test name, such as:
com.ibm.icu.dev.test.calendar.CompatibilityTest.

The tests that we want to be run are specified manually in
CtsIcuTests.xml which is copied over to the test repository
which cts-tradefed looks for test definitions in. This is
based on the functionality of ICU which is used within libcore.

If a package name is not specified, all the tests that would
normally be run by the ICU test runner, will be run.

Results are reported individually back to the instrumentation
to allow the host to determine how many tests failed / passed.

The tests appear in the cts-tradefed console as the package
com.ibm.icu.dev.test. One can run all the tests which have been
whitelisted with:

run cts --package com.ibm.icu.dev.test

or a specific class with:

run cts --class com.ibm.icu.dev.test.calendar.CalendarTest

Change-Id: I847525918525ceab7a276b143271853a2c7e9465