Check that secure lock screen is not set up.
CTS test AndroidKeyStoreTest#testKeyStore_SetEntry_PrivateKeyEntry_Params_Unencrypted_Failure
assumes that the secure lock screen is not set up but does not check
that. As a result, it fails with a cryptic error when the secure lock
screen is set up.

This change addresses the issue by: (1) documenting this assumption in
the test, and (2) making the test check this assumption and fail with
a much friendlier error if the assumption is violated.

Change-Id: I7a09c83b93f52fe43884e26ba0795480c44ec6a3