Add CTS tests for suggested print attributes
It is possible for a app to suggest print attributes. E.g. a word
processor might suggest the same paper size as already selected in the
app. If an attribute is not suggested or the suggestion could not be met
by the printer the default value for this attribute should be chosen.

Change-Id: I11f8417945344af28497ab772778377315f9a218