Adjust Refocus test to allow minor diffs in pixels
Bug: 26087547

Currently, if more than 0.01% pixels are different from the
reference image, the test will fail, although the average
difference can be tiny.

Adjust the test to check the average difference as well, and only
report failure if the average diff is greater than 5 (out of 255),
in addition to 0.01% in the total number of different pixels.

Change-Id: Ie909d0c014e73e5bc802b81ad2c47433a1465ccf