CTS tests for single-source RS allocation API
Bug: 23535985

Adapted from Pirama's tests in RsTest, to make it work with the CTS
testing harness.

Refactored so that tests are in seperate groups depending on what
they are testing. In the allocation creation test, also moved the loop
over data types and vector sizes into the script from Java.

Dropped the return values for testing kernels, since it is no longer
a requirement in the final implementation of rsForEach().

Change-Id: Ie5512809412c1229dd975395fd33c18ec643e1dc