EncryptionTest: add "xts-aes-aesni"
The EncryptionTest#testConfig method was failing when checking
for valid crypto routines in the kernel as reported by
/proc/crypto. Intel  architectures can leverage the AES Native
Instruction Set functionality as a replacement to the ARM v8
Cryptography Extensions (CE).
To provide a work-around to this ARMv8 CTS test case failure,
we add the intel AES driver name to the white-list of allowed
configurations.

Change-Id: Id7c5734438f48d90b7464f47bff0e14dddc54df6