Fix MockInCallService's default starting values
mIsServiceUnbound in MockInCallService used to have no initial value,
which means that it defaulted to false, effectively meaning that
MockInCallService started out thinking that it was bound. This caused
test failures in tests that did not make use of MockInCallService
because the cleanup code asserted that MockInCallService was unbound.

This change renames mIsServiceUnbound to mIsServiceBound and explicitly
sets the initial value to false.

Bug: 26641212
Change-Id: I03c6fb0b994c4b55bf55f9dbedd85fd3add8ef5e