Avoid long delay between running CtsIcuTestCases and receiving results
Fixes IcuTestRunner so it handles the ARGUMENT_TEST_CLASS
argument correctly.

Handles the log argument properly by sending the CTS tradefed
shell the events that it would send if it ran all the tests
without actually running the tests. This avoids a long (around
10 minute) delay between the request to the shell to run the
tests and any results being returned.

Change-Id: Ie24dc65348906042a44995f8a51387d478e2852a