Update manifest snapshot against which permissions are verified.
We have a copy of the Android manifest which is used as a golden
version against which we compare the manifest of the tested device
to catch undesirable permission changes made by a partner. We added
a few permissions lately and this change updates the golden manifest
against which we benchmark.

bug:27240052

Change-Id: I98090247722d99292589145669842b4fe77d881f