CameraITS: add lens shading and color uniformity test
Test that the lens shading correction is applied appropriately,
and color of a monochrome uniform scene is evenly distributed,
for example, when a diffuser is placed in front of the camera.
Perform this test on a yuv frame with auto 3a. Lens shading is
evaluated based on the y channel. Measure the average y value
for each sample block specified, and then determine pass/fail
by comparing with the center y value. The color uniformity test
is evaluated in r/g and b/g space. At specified radius of the
image, the variance of r/g and b/g value need to be less than
a threshold in order to pass the test.

Change-Id: I6cf7676ef7f3ec8ef9a8455675ee212287637baf