sensor test: Fix PROXIMITY FIFO test.
In version 6.0 of the CCD, the last sentance of section 7.3.9
states that if a SENSOR_TYPE_PROXIMITY is present, then it must
have a minimum buffering of 100 events.

Our previous test, however, was checking that our proximity sensor
would have an exclusive buffer of 300 or _less_ dedicated to
proximity events.  That's wrong in two ways, as it will fail
hardware that can buffer 512 of these events, and pass hardware
that can only buffer up to 64 of these events.

We change this test to use the constant from the CCD, and to
correctly test against it as a minimum.

Change-Id: I4db062869125a0de0b255fa8f8b8217e90718c83