Stop ignoring untriaged ICU CTS test failures
This stops filtering tests that are known to fail but which have
not yet been investigated out of the set of tests to run. The
failing tests are still listed in the icu-knownn-failures.txt
file but are marked as being expected to succeed so they are
still run. That gives us a convenient list of the failures and
makes it easier to run just the failing tests.

It also updates the list of tests for old-cts to run to include
the failing tests.

Bug: 22023363
Change-Id: I2eda143182f65b3c144805db220c9587d552f866