Remove wifi MAC addr access test
The Openjdk version of NetworkInterface.java obtains the wifi MAC
address using a netlink route socket, instead of the previous method
of finding it in /sys.

Since any app can create and use a netlink route socket to get this
information, specifically blocking it in NetworkInterface is not
useful. Delete test and find a better way to block access.

Bug: 27266415
Change-Id: Ifb4ec7cfd565d726390a535e247b20e556f834e6