android.keystore.cts: move KeyStoreTest from libcore
Need to increase the timeout for a test and the Makefile target
we need to add as to set the timeout (ctsdeviceutil) cannot be used
in libcore tests.

Anyway, this is more of an integration test that can live here with
high-level tests.

The timeout is needed as the test test_KeyStore_Builder is taking too
long on armeabi-v7a and making test bots unhappy.
Filed b/27810271 to check what is going on.

Bug: 27806958
Bug: 27810271
Change-Id: Ifd4c6debbd7d27c0c3e619a794ccff3429a0182d