Add global enforcing test.
SELinux tests are primarily enforced via host-side tests to take
advantage of privileges available to the shell user that are not
available to cts.  Add a check that explicitly checks to make sure
SELinux policy is being enforced on the system, rather than relying on
results from various checkAccess() tests in SELinuxHostTest.

Bug: 27756382
Bug: 27763829
Change-Id: Ic7e94267ef1ad23491539e12d960014b0440872a