Camera2: PerformaceTest#testSingleCapture: Wait for preview to stop
stopPreview would simply close the session asynchronously, and it was
possible for frames to still be produced when the next iteration started
configuring the device.

Configuration would start by closing the previous imagereader, at which
point the LEGACY layer could have an invalid Surface it is drawing to.
This results in a fatal device error for LEGACY currently.

Instead, actually wait for preview to drain before closing off readers.

Bug: 27828605
Change-Id: Ifd330d7c8da7c7cf1ed33b78c493b015eee643a9