OpenGlEsVersionTest: require features to exactly match capabilities
Previously, a device was allowed to provide a higher OpenGL ES version
than it claimed via PackageManager system features. Similarly, it
could support the AEP extension without claiming the AEP system
feature.

This led to under-reporting of capabilities as OEMs forgot to set the
feature properly (see b/26984293 for example), and prevented
developers from successfully targetting devices with the desired
features.

The test now requires that the system features reflect the actual
OpenGL ES version supported and the presence of the AEP extension.

Change-Id: Ife45d872965f8c8dd1d9f9a64c2e28aad9886769