Disable IME for android.text.method.cts tests
Changes in IME affects the tests that use KeyEvents. This CL makes tests
to send KeyEvents directly to the InputMethodManager of the test app,
instead of relying on IME. Moreover, it uses TextView and EditText
variants that return null for onCreateInputConnection to minimize the
effects of the keyboard on the tests.

Bug: 28352584
Bug: 28346955
Change-Id: Ied227ff5573146a873e3ded70ec22ed1b063386c