Fix accessibility gesture with magnfification test
The test was verifying behavior it didn't really care about - whether
or not the viewport moved to make the magnified region include the
upper right corner of the screen. It was also requesting (0,0) for
the center point, which can be invalid if there is a magnification
border. Re-centering on the expected correct center.

Bug: 28460665
Change-Id: Ib9cabe862ed59fc7dafb2b3ecb725decb750c520