Fixed NoActivityRelatedPermissionTest failure
The test was previously checking if there was 2 tasks
(1 for the current activity and the other for the home activity)
returned for the getRecentTasks call, however it is possible for
us to have multiple home tasks as some point and the test will
fail. We now filter out all the home tasks from the returned list
before checking that there is 1 or less task in the results.

Bug: 28552936
Change-Id: Ife7db8a9643e2cb7070172b0030ee2bc7063a62f