common: Have "assume" failures PASS
When we fail a JUnit4 "assume" check, we do not want to report
NOT_EXECUTED, as that indicates to the user that there was a
problem trying to run the test, and we need to run it again.

Eventually (b/28386054), we want to report this as SKIP to the
user.  So we introduce some partial plumbing to allow that in
the future.  But this will require a change of the TestStatus
enum, which is out of scope for this fix.  So for now, we use
PASS so the user knows this isn't a CTS/GTS issue.

Bug: 28633728
Change-Id: Idf370aa7c44f1343d413a6301edcd02965bdd478