Fix for Rotation Vector CV CrossCheck test
The test case is implemented in a way that the surface view size adjustment
will be completed in two seconds. However, the size adjustment routine does not allow
the aspect ratio torlerance and also does not consider the precision lost during the
'float' to 'int' type casting, the size adjustment of a surface view can be
executed for several times.

The FOV value of camera parameter will be changed according to the picture size.
The 'wrong' FOV value based on a previous picture size will be delivered to the
analyzer of testcase and the testcase will fail during the analyze phase.
The cts verifier application need to be changed to acquire a proper FOV value
 from the camera device after configuring the picture size.

b/28153914

Change-Id: Ic83553f2df1e2006ea4d74b764cf60a876ed8e13
Signed-off-by: Raj Mamadgi <r.mamadgi@samsung.com>