RVCVXCheck - avoid loading OpenCV from play store
If OpenCV Manager is not installed by the time of test, by default
the OpenCV loader prompt user to go to play store to install.
The version at OpenCV website can change over time and becomes
incompatible. This CL avoid this behavior and notify user to install
OpenCV Manager manually if it does not exist in system.

Bug: 28849922

Change-Id: Ic83edd02d0759715a63aa66e9595723d7230eaa1