Consider only platform defined groups
In the test that verifies if a partner modified the platform
permission policy we were also considering permission groups
defined by apps. While apps should not be declaring groups in
the platform namespace, the scope of this test does not include
verifying this. Instead, apps should be prevented from declaring
groups and permissions in the platform namespace at the OS level.

bug:29532353

Change-Id: Ic8d76c16dc183f7930e2748330fac30abed1b00b