Add a host-side test to verify /proc/[PID]/task/[TID]/stat exists.
Android Studio Performance Tools needs, as the shell user, to read
the files to collect live metrics data. It requires the files are in
a reasonable shape.

I have verified this test passing on a Nexus 6P running master branch.
I have also verified this test passing on a non-Nexus device (Samsung
S7 running 6.0.1). In order to run the test on the non-Nexus device,
I have ported this CL into mnc-dev branch (http://ag/1390776) which
will not be submitted. I will only submit this CL on master branch.

http://b/31183357 is tracking this work.

Change-Id: I8637481d72b3e2d2ebb1441f63d71a1e10c679a5