Tests: Fix focus problems when using optimized fragment transactions.
Bug 33754255

Instead of making Views INVISIBLE immediately after adding them,
Views are set to transitionAlpha = 0.

This CL modifies the tests to check that the view is visible in
a way that can track transitionAlpha. It also adds a test to ensure
that the focus works after adding a fragment that grabs focus
in a transaction.

Test: ran fragment CTS tests

Change-Id: I966da3051912fff0f7ac42dec31626fd8248f7c2