Set retry commandline args for all CTS shards
Each shard has its own instance of BuildInfo, so prior to this change,
only one shard is informed of the commandline arguments used in the session
that will be retried. This shard may or may not be the master shard, which
writes the result report. To ensure that the master shard has the correct retry
args info (and thus writes the correct args to the result, which is important
for further retries), set this attribute across all shards.

bug:33211104
bug:33380715
Test: run collect-tests-only -r 0 --shards 2
Change-Id: Iadf567c383cdbc07f1460bf9639b3a569f041ee9