CTS tests for new storage/quota APIs.
Consists of a single host app that drives three separate apps
which can be asked to allocate disk space in different ways.  The
final app holds the PACKAGE_USAGE_STATS permission, so it can query
stats for the first two packages.

Each package uses Utils.useSpace() which allocates blobs of various
types in different storage locations, all of which we expect to be
tracked correctly by the OS.  The size of each blob is a prime number
to aid debugging.  Test writing blobs both manually and using
fallocate().

Start testing cache behavior flags, which will be finished in a
future CL once we have better adb commands for control.

Test: cts-tradefed run commandAndExit cts-dev -m CtsAppSecurityHostTestCases -t android.appsecurity.cts.StorageHostTest
Bug: 27948817, 30927649, 33811826
Change-Id: Ie6bd172f02c1ae17f6b4e33acddc43bc693e55f2