Prepare for removal of junit and android.test classes from Android API (step 1)
In preparation for removing junit and android.test classes from
the Android API this change adds explicit dependencies on junit
and/or legacy-android-test to ensure that modules will compile
properly once it is removed.

Care was taken to ensure that any targets that depended,
directly or indirectly, on android-support-test did not also
depend on junit or hamcrest as they conflict with the versions
embedded within android-support-test.

Bug: 30188076
Test: make checkbuild
Merged-In: Ia67b0c29bfcaf36bb1ec4e4c97f5aa5d0a0b3b22
Change-Id: Ib0f1ff24f4190e0d7b3743b984994431ac6dcc91