DO NOT MERGE: Allow not executing "Network Connectivity Screen Off Test".
This test is trying to validate the IPv6 reliability requirement of
the CDD 7.4.5, more specifically that unicast IPv6 packets sent to the
device MUST NOT be dropped, even when the screen is not in an active
state.

Some devices might have no way to turn off the screen, which is by
itself not a compatibility violation, but the test would not be able
to be executed.

Bug: 34226510
Test: None (no change in behavior by this CL.)
Change-Id: Ia4e8033736dae875c5467311aaace1af3d1366a3