Disable SelfManagedConnectionServiceTest#testEmergencyCallOngoing.
Disable the SelfManagedConnectionServiceTest#testEmergencyCallOngoing
CTS test; it is relying on setting the RIL ECC List, which apparently
can't be done through a shell command any more.

This test is ultimately not required for completeness of the API, so I'm
going to disable it for now until the next release when we can add a
test API To support setting test emergency numbers during CTS runs.

Test: CTS
Bug: 36999709
Change-Id: I9d073c6c846a244f430376b293ef5ebe17703ba7